Split webview message handler into modules (oh-tab-mmk)
