+++ src/webview/host/createWebviewMessageHandler.ts
@@ -1,149 +1,29 @@
 import * as vscode from 'vscode';
-import * as fs from 'fs/promises';
-import * as path from 'path';
-import * as os from 'os';
-import * as childProcess from 'child_process';
-import { assertValidProfileId, DEFAULT_PROVIDER_BASE_URLS, detectProviderFromBaseUrl, LLMProfileValidationError, type LLMProvider, type SecretRegistry } from '@openhands/agent-sdk-ts';
+import type { SecretRegistry } from '@openhands/agent-sdk-ts';
 import { SettingsManager } from '../../settings/SettingsManager';
 import { VscodeSettingsAdapter } from '../../settings/VscodeSettingsAdapter';
-import { ElevenLabsTtsService } from '../../hal/elevenlabs/ttsService';
-import { TtsConversationGate } from '../../hal/elevenlabs/ttsConversationGate';
-import { classifyHalVoiceDecision } from '../../hal/gemini/decisionClassifier';
-import { getHalDialogueLinesForMode } from '../../shared/halScript';
-import { DEFAULT_HAL_LLM_PROFILE_ID } from '../../shared/halDefaults';
 import { resolveConfiguredLlmLabel } from '../../shared/llmProfiles';
-import { OPENHANDS_IMAGE_URL_PREFIX, getGlobalStorageBaseDir, getPastedImagePath, parseBase64DataImageUrl, rewriteDataImageMarkdown, rewriteOpenHandsImageUrls } from '../../shared/pastedImages';
-import { MAX_PASTED_IMAGE_BYTES } from '../../shared/pasteLimits';
-import { normalizeServerUrl } from '../../shared/serverUrls';
-import { STATUS_MESSAGE_DISMISS_DELAY_MS, type HostToWebviewMessage, type WebviewToHostMessage } from '../../shared/webviewMessages';
-import { buildAttachmentBlocks, safeParseUri, toAttachmentLabel } from './attachments';
+
+import type { HostToWebviewMessage, WebviewToHostMessage } from '../../shared/webviewMessages';
 // Environment info is provided via AgentContext.userMessageSuffix (extension host).
-import { getEffectiveWorkspaceRoot, resolvePreferredWorkspaceFolderUri } from '../../shared/workspaceRoot';
-import { getConversationHistoryList, persistConversationTitle } from './conversationHistory';
-import { showWorkspaceDiff } from './diffDocuments';
-import { resolveGitHeadDiffContents } from './gitHeadDiff';
-import * as llmProfilesStore from './llmProfilesStore';
+
 import { listSkillFiles } from './skills';
 import { listWorkspaceFiles } from './workspaceFiles';
-import { resolveWorkspaceFilePath } from './workspacePaths';
-import { getDefaultLocalToolIds, listLocalToolDescriptors, normalizeLocalToolIds, resolveLocalTools, type LocalToolId } from '../../shared/localTools';
-import { summarizeWithLocalLlm } from '../../extension/summarizeWithLocalLlm';
+import { getDefaultLocalToolIds, resolveLocalTools } from '../../shared/localTools';
+import { handleWebviewConsole, handleWebviewError, handleWebviewNetwork, handleWebviewWebSocket } from './handlers/devBridge';
+import { handleHalStateResponse, handleRenderedEventsResponse, handleUiStateResponse } from './handlers/stateResponses';
+import { handleOpenMarkdownLink, handleOpenSkill, handleOpenWorkspaceDiff, handleOpenWorkspaceFile } from './handlers/openers';
+import { handleOpenAttachment, handleSelectAttachments } from './handlers/attachments';
+import { createPostToolsList, handleSetEnabledTools, isLocalConversationToolControls } from './handlers/tools';
+import { handleDeleteConversation, handleRequestHistory, handleRestoreConversation } from './handlers/history';
+import { handleAddServer, handleRemoveServer, handleSelectServer, handleSwitchToLocal } from './handlers/servers';
+import { handleSend } from './handlers/send';
+import { computeWelcomeSecretStatus, handleLlmProfileApiKeySetRequest, handleLlmProfileApiKeyStatusRequest, handleLlmProfileDeleteRequest, handleLlmProfileLoadRequest, handleLlmProfileSaveRequest, handleLlmProfilesListRequest, handleSetLlmProfileId, listAvailableLlmProfiles } from './handlers/llmProfiles';
+import { createElevenlabsTtsGateFactory, handleHalTtsRequest, handleHalVoiceConfirmRequest } from './handlers/hal';
 
 export type WebviewHost = {
   postMessage: (message: HostToWebviewMessage) => Thenable<boolean>;
 };
-
-type LocalConversationToolControls = {
-  mode: 'local';
-  getToolNames: () => string[];
-  setTools: (tools: unknown[]) => void;
-};
-
-const isLocalConversationToolControls = (value: unknown): value is LocalConversationToolControls => {
-  if (!value || typeof value !== 'object') return false;
-  const candidate = value as Partial<LocalConversationToolControls> & { [k: string]: unknown };
-  return candidate.mode === 'local'
-    && typeof candidate.getToolNames === 'function'
-    && typeof candidate.setTools === 'function';
-};
-
-function execFileText(command: string, args: string[], cwd?: string): Promise<string> {
-  return new Promise((resolve, reject) => {
-    childProcess.execFile(command, args, { cwd }, (err, stdout, stderr) => {
-      if (err) {
-        const message = typeof stderr === 'string' && stderr.trim().length > 0 ? stderr.trim() : err.message;
-        reject(new Error(message));
-        return;
-      }
-      resolve(typeof stdout === 'string' ? stdout : String(stdout));
-    });
-  });
-}
-
-function normalizeGeneratedConversationTitle(raw: string): string | undefined {
-  const firstLine = raw
-    .split('\n')
-    .map((line) => line.trim())
-    .find((line) => line.length > 0);
-  if (!firstLine) return undefined;
-
-  const unquoted = firstLine.replace(/^["'`]+/, '').replace(/["'`]+$/, '').trim();
-  if (!unquoted) return undefined;
-
-  const strippedPrefix = unquoted.replace(/^title\\s*:\\s*/i, '').trim();
-  if (!strippedPrefix) return undefined;
-
-  const words = strippedPrefix.split(/\\s+/).filter(Boolean);
-  if (words.length === 0) return undefined;
-  return words.slice(0, 7).join(' ');
-}
-
-async function persistPastedImage(baseDir: string, imageId: string, bytes: Uint8Array): Promise<void> {
-  const filePath = getPastedImagePath(baseDir, imageId);
-  await fs.mkdir(path.dirname(filePath), { recursive: true });
-  await fs.writeFile(filePath, bytes);
-}
-
-const ensureRequiredLocalToolIds = (toolIds: LocalToolId[]): LocalToolId[] => {
-  // `finish` is always enabled by the runtime agent (for safe termination).
-  // Keep the UI/tool-selection source of truth consistent with what is actually sent to the LLM.
-  if (toolIds.includes('finish')) return toolIds;
-  return [...toolIds, 'finish'];
-};
-
-type RemoteToolListDeps = {
-  serverUrl: string;
-  sessionApiKey?: string;
-};
-
-const getRemoteToolListDeps = (conversation: unknown): RemoteToolListDeps | null => {
-  if (!conversation || typeof conversation !== 'object') return null;
-  const candidate = conversation as { [k: string]: unknown };
-  const serverUrl = candidate.serverUrl;
-  if (typeof serverUrl !== 'string' || serverUrl.trim().length === 0) return null;
-
-  const rawSessionKey =
-    (candidate.settings as { secrets?: { sessionApiKey?: unknown } } | undefined)?.secrets?.sessionApiKey;
-  const sessionApiKey = typeof rawSessionKey === 'string' && rawSessionKey.trim().length > 0 ? rawSessionKey : undefined;
-  return { serverUrl: serverUrl.trim(), sessionApiKey };
-};
-
-async function fetchRemoteToolNames(params: RemoteToolListDeps): Promise<string[] | null> {
-  const normalized = normalizeServerUrl(params.serverUrl);
-  if (!normalized.ok) return null;
-  const url = `${normalized.url}/api/tools/`;
-  const headers: Record<string, string> = {};
-  if (params.sessionApiKey) {
-    headers['X-Session-API-Key'] = params.sessionApiKey;
-  }
-
-  const fetchFn = globalThis.fetch;
-  if (typeof fetchFn !== 'function') return null;
-
-  const abortController = new AbortController();
-  const timeoutMs = 4000;
-  const timeout = setTimeout(() => abortController.abort(), timeoutMs);
-
-  try {
-    const res = await fetchFn(url, { method: 'GET', headers, signal: abortController.signal });
-    if (!res.ok) return null;
-    const payload = await res.json();
-    if (!Array.isArray(payload)) return null;
-
-    const out: string[] = [];
-    for (const tool of payload) {
-      if (typeof tool !== 'string') continue;
-      const trimmed = tool.trim();
-      if (!trimmed) continue;
-      out.push(trimmed);
-    }
-    return out;
-  } catch {
-    return null;
-  } finally {
-    clearTimeout(timeout);
-  }
-}
 
 export type CreateWebviewMessageHandlerDeps = {
   context: vscode.ExtensionContext;
@@ -219,33 +99,10 @@ export type CreateWebviewMessageHandlerDeps = {
 export function createWebviewMessageHandler(deps: CreateWebviewMessageHandlerDeps) {
   const { context, host } = deps;
   const settingsMgr = new SettingsManager(new VscodeSettingsAdapter(context));
-  const elevenlabsCacheMaxBytes = 50 * 1024 * 1024;
-  let elevenlabsTtsGate: TtsConversationGate | null = null;
+  const getElevenlabsTtsGate = createElevenlabsTtsGateFactory({ context, maxCacheBytes: 50 * 1024 * 1024 });
   const historyTitleGenerationInFlight = new Set<string>();
 
-  const postToolsList = async (): Promise<void> => {
-    const mode = deps.getConversationMode();
-    if (mode !== 'local') {
-      const conversation = deps.getConversation();
-      const remoteDeps = getRemoteToolListDeps(conversation);
-      const toolNames = remoteDeps ? await fetchRemoteToolNames(remoteDeps) : null;
-      const tools = (toolNames ?? []).map((name) => ({ id: name, label: name }));
-      void host.postMessage({ type: 'toolsList', tools, enabledToolIds: toolNames ?? [] });
-      return;
-    }
-
-    const tools = listLocalToolDescriptors();
-    const conversation = deps.getConversation();
-    const enabledToolIds = (() => {
-      if (isLocalConversationToolControls(conversation)) {
-        const ids = normalizeLocalToolIds(conversation.getToolNames());
-        if (ids !== null) return ids;
-      }
-      return getDefaultLocalToolIds();
-    })();
-
-    void host.postMessage({ type: 'toolsList', tools, enabledToolIds: ensureRequiredLocalToolIds(enabledToolIds) });
-  };
+  const postToolsList = createPostToolsList({ deps, host });
 
   const postStatusError = (message: string): void => {
     void host.postMessage({
@@ -257,152 +114,12 @@ export function createWebviewMessageHandler(deps: CreateWebviewMessageHandlerDep
     });
   };
 
-  const validateProfileId = (profileId: string): void => {
-    try {
-      assertValidProfileId(profileId);
-    } catch (err) {
-      if (err instanceof LLMProfileValidationError) {
-        throw new Error(err.message);
-      }
-      throw err;
-    }
-  };
-
-  const getProfileApiKeySecretKey = (profileId: string): string => {
-    validateProfileId(profileId);
-    return `openhands.llmProfileApiKey.${profileId}`;
-  };
-
-  const getProviderApiKeyName = (provider: LLMProvider): string => {
-    switch (provider) {
-      case 'openrouter':
-        return 'OPENROUTER_API_KEY';
-      case 'litellm_proxy':
-        return 'LITELLM_API_KEY';
-      case 'anthropic':
-        return 'ANTHROPIC_API_KEY';
-      case 'gemini':
-        return 'GEMINI_API_KEY';
-      default:
-        return 'OPENAI_API_KEY';
-    }
-  };
-
-  const isLlmProvider = (value: string): value is LLMProvider => {
-    return value === 'openai'
-      || value === 'litellm_proxy'
-      || value === 'openrouter'
-      || value === 'anthropic'
-      || value === 'gemini';
-  };
-
-  const getStoredSecret = async (key: string): Promise<string | undefined> => {
-    const trimmedKey = key.trim();
-    if (!trimmedKey) return undefined;
-    try {
-      const resolved = deps.secretRegistry
-        ? await deps.secretRegistry.get(trimmedKey)
-        : (process.env[trimmedKey] ?? (await context.secrets.get(trimmedKey)));
-      const trimmedValue = typeof resolved === 'string' ? resolved.trim() : '';
-      return trimmedValue || undefined;
-    } catch {
-      return undefined;
-    }
-  };
-
-  const hasStoredSecret = async (key: string): Promise<boolean> => {
-    return Boolean(await getStoredSecret(key));
-  };
-
-  const getElevenlabsTtsGate = (): TtsConversationGate => {
-    if (elevenlabsTtsGate) return elevenlabsTtsGate;
-    const baseDir = context.globalStorageUri?.fsPath || path.join(os.tmpdir(), 'oh-tab-global-storage');
-    const cacheDir = path.join(baseDir, 'hal', 'elevenlabs', 'tts-cache');
-    elevenlabsTtsGate = new TtsConversationGate(
-      new ElevenLabsTtsService({
-        cacheDir,
-        maxCacheBytes: elevenlabsCacheMaxBytes,
-      })
-    );
-    return elevenlabsTtsGate;
-  };
-
   return async (msg: unknown) => {
     if (!msg || typeof msg !== 'object' || !('type' in msg)) return;
     const message = msg as WebviewToHostMessage;
 
     const outputChannel = deps.getOutputChannel();
     const conversation = deps.getConversation();
-
-    const llmProfileStoreOptions = (): { rootDir?: string } => {
-      const rootDir = typeof deps.getLlmProfilesStoreRoot === 'function' ? deps.getLlmProfilesStoreRoot() : undefined;
-      if (typeof rootDir !== 'string') return {};
-      const trimmed = rootDir.trim();
-      return trimmed ? { rootDir: trimmed } : {};
-    };
-
-    const listAvailableLlmProfiles = (): string[] => {
-      try {
-        return llmProfilesStore.listProfiles(llmProfileStoreOptions());
-      } catch (err) {
-        const reason = err instanceof Error ? err.message : String(err);
-        outputChannel?.appendLine(`[llm] Failed to list profiles: ${reason}`);
-        return [];
-      }
-    };
-
-    const sendHistoryList = async (): Promise<void> => {
-      try {
-        const convRoot = deps.getConversationStoreRoot() ?? (await deps.resolveConversationStoreRoot());
-        const conversations = await getConversationHistoryList(convRoot, outputChannel);
-        void host.postMessage({ type: 'historyList', conversations });
-
-        // Best-effort: generate short titles for items that don't have one persisted yet.
-        // Non-blocking: HistoryView should render immediately and then update as titles become available.
-        void (async () => {
-          const secrets = deps.secretRegistry;
-          if (!secrets) return;
-
-          const missing = conversations
-            .filter((c) => !c.title && typeof c.firstMessage === 'string' && c.firstMessage.trim().length > 0)
-            .sort((a, b) => b.timestamp - a.timestamp)
-            .slice(0, 30);
-          if (missing.length === 0) return;
-
-          const settings = await settingsMgr.get();
-
-          for (const convo of missing) {
-            if (historyTitleGenerationInFlight.has(convo.id)) continue;
-            historyTitleGenerationInFlight.add(convo.id);
-            try {
-              const prompt =
-                `Generate a short conversation title (max 7 words).\\n` +
-                `Return ONLY the title text (no quotes, no punctuation at the end).\\n\\n` +
-                `First user message:\\n${convo.firstMessage}`;
-
-              const raw = await summarizeWithLocalLlm(settings, prompt, secrets);
-              const title = normalizeGeneratedConversationTitle(raw);
-              if (!title) continue;
-
-              await persistConversationTitle(convRoot, convo.id, title, outputChannel);
-              convo.title = title;
-              void host.postMessage({ type: 'historyList', conversations });
-            } catch (err) {
-              const reason = err instanceof Error ? err.message : String(err);
-              outputChannel?.appendLine(`[history] Failed to generate title for ${convo.id}: ${reason}`);
-              // If this fails once (missing key/profile/etc.), avoid spamming more calls this round.
-              break;
-            } finally {
-              historyTitleGenerationInFlight.delete(convo.id);
-            }
-          }
-        })();
-      } catch (err) {
-        const reason = err instanceof Error ? err.message : String(err);
-        outputChannel?.appendLine(`[history] ${reason}`);
-        void host.postMessage({ type: 'historyList', conversations: [] });
-      }
-    };
 
     switch (message.type) {
       case 'webviewReady': {
@@ -425,7 +142,7 @@ export function createWebviewMessageHandler(deps: CreateWebviewMessageHandlerDep
 
         void host.postMessage({
           type: 'llmProfilesUpdated',
-          profiles: listAvailableLlmProfiles(),
+          profiles: listAvailableLlmProfiles({ deps, outputChannel }),
           activeProfileId: initSettings.llm.profileId ?? null,
         });
 
@@ -444,54 +161,8 @@ export function createWebviewMessageHandler(deps: CreateWebviewMessageHandlerDep
 
         // Welcome page: communicate whether API keys are present so the UI can show onboarding prompts.
         void (async () => {
-          const secretIsSet = async (key: string): Promise<boolean> => {
-            try {
-              const value = await context.secrets.get(key);
-              return typeof value === 'string' && value.trim().length > 0;
-            } catch {
-              return false;
-            }
-          };
-
-          const envIsSet = (key: string): boolean => {
-            const value = process.env[key];
-            return typeof value === 'string' && value.trim().length > 0;
-          };
-
-          const hasGeminiKey = (await secretIsSet('GEMINI_API_KEY'))
-            || envIsSet('GEMINI_API_KEY')
-            || (initSettings.llm.provider === 'gemini' && typeof initSettings.secrets.llmApiKey === 'string' && initSettings.secrets.llmApiKey.trim().length > 0);
-
-          const hasGenericKey = typeof initSettings.secrets.llmApiKey === 'string' && initSettings.secrets.llmApiKey.trim().length > 0;
-
-          const providerStorageKeys = [
-            'OPENAI_API_KEY',
-            'ANTHROPIC_API_KEY',
-            'OPENROUTER_API_KEY',
-            'LITELLM_API_KEY',
-          ];
-          let hasAnyProviderStorageKey = false;
-          for (const key of providerStorageKeys) {
-            if (envIsSet(key) || await secretIsSet(key)) {
-              hasAnyProviderStorageKey = true;
-              break;
-            }
-          }
-
-          let hasAnyProfileKey = false;
-          try {
-            for (const profileId of listAvailableLlmProfiles()) {
-              if (await secretIsSet(`openhands.llmProfileApiKey.${profileId}`)) {
-                hasAnyProfileKey = true;
-                break;
-              }
-            }
-          } catch {
-            hasAnyProfileKey = false;
-          }
-
-          const hasProviderKey = hasGeminiKey || hasGenericKey || hasAnyProviderStorageKey || hasAnyProfileKey;
-          void host.postMessage({ type: 'welcomeSecretStatus', hasProviderKey, hasGeminiKey });
+          const status = await computeWelcomeSecretStatus({ deps, context, settings: initSettings });
+          void host.postMessage({ type: 'welcomeSecretStatus', hasProviderKey: status.hasProviderKey, hasGeminiKey: status.hasGeminiKey });
         })();
 
         deps.flushConversationEventBacklog({
@@ -525,190 +196,36 @@ export function createWebviewMessageHandler(deps: CreateWebviewMessageHandlerDep
         break;
       }
       case 'setEnabledTools': {
-        const mode = deps.getConversationMode();
-        if (mode !== 'local') break;
-
-        const normalized = normalizeLocalToolIds(message.toolIds);
-        if (normalized === null) {
-          postStatusError('Invalid tools selection received from webview');
-          break;
-        }
-
-        const conversation = deps.getConversation();
-        if (!isLocalConversationToolControls(conversation)) {
-          postStatusError('Cannot update tools: conversation unavailable');
-          break;
-        }
-
-        const toolIds = ensureRequiredLocalToolIds(normalized);
-        try {
-          conversation.setTools(resolveLocalTools(toolIds));
-        } catch (err) {
-          const reason = err instanceof Error ? err.message : String(err);
-          outputChannel?.appendLine(`[tools] Failed to update tools: ${reason}`);
-          void host.postMessage({ type: 'statusMessage', level: 'info', message: reason, autoDismiss: true, autoDismissDelay: 4000 });
-        }
-
-        await postToolsList();
+        await handleSetEnabledTools({ deps, host, outputChannel, postStatusError, postToolsList, message });
         break;
       }
       case 'openSkill': {
-        const skillPath = message.path;
-        if (!skillPath) break;
-        try {
-          const skillsRoot = path.resolve(os.homedir(), '.openhands', 'skills');
-          const resolvedPath = path.resolve(skillPath);
-          const relative = path.relative(skillsRoot, resolvedPath);
-          if (relative.startsWith('..') || path.isAbsolute(relative)) {
-            void vscode.window.showErrorMessage('Refusing to open skill outside of ~/.openhands/skills');
-            break;
-          }
-          const document = await vscode.workspace.openTextDocument(vscode.Uri.file(resolvedPath));
-          await vscode.window.showTextDocument(document, { preview: false });
-        } catch (err) {
-          const reason = err instanceof Error ? err.message : String(err);
-          void vscode.window.showErrorMessage(`Failed to open skill file: ${reason}`);
-        }
+        await handleOpenSkill(message);
         break;
       }
       case 'openWorkspaceFile': {
-        const p = message.path;
-        if (!p) break;
-        try {
-          const { resolvedPath } = resolveWorkspaceFilePath(p);
-          await fs.stat(resolvedPath);
-          const document = await vscode.workspace.openTextDocument(vscode.Uri.file(resolvedPath));
-          await vscode.window.showTextDocument(document, { preview: false });
-        } catch (err) {
-          const reason = err instanceof Error ? err.message : String(err);
-          void vscode.window.showErrorMessage(`Failed to open file: ${reason}`);
-        }
+        await handleOpenWorkspaceFile(message);
         break;
       }
       case 'openMarkdownLink': {
-        const raw = typeof message.href === 'string' ? message.href.trim() : '';
-        if (!raw || raw.startsWith('#')) break;
-
-        // Only allow http(s)/mailto links and workspace-internal file links.
-        if (/^https?:\/\//i.test(raw) || /^mailto:/i.test(raw)) {
-          const uri = safeParseUri(raw);
-          if (!uri || (uri.scheme !== 'http' && uri.scheme !== 'https' && uri.scheme !== 'mailto')) {
-            void vscode.window.showErrorMessage('Blocked unsafe link.');
-            break;
-          }
-          await vscode.env.openExternal(uri);
-          break;
-        }
-
-        const wsRoot = getEffectiveWorkspaceRoot();
-        if (!wsRoot) {
-          void vscode.window.showErrorMessage('Cannot open link: no workspace folder is open.');
-          break;
-        }
-
-        const withoutFragment = raw.split('#')[0];
-        const withoutQuery = withoutFragment.split('?')[0];
-        const inputPath = withoutQuery.trim();
-        if (!inputPath) break;
-
-        const resolvedPath = path.isAbsolute(inputPath) ? path.resolve(inputPath) : path.resolve(wsRoot, inputPath);
-        const rel = path.relative(wsRoot, resolvedPath);
-        const inWorkspace = rel && !rel.startsWith('..') && !path.isAbsolute(rel);
-        if (!inWorkspace) {
-          void vscode.window.showErrorMessage('Blocked unsafe link.');
-          break;
-        }
-
-        try {
-          await fs.stat(resolvedPath);
-          const document = await vscode.workspace.openTextDocument(vscode.Uri.file(resolvedPath));
-          await vscode.window.showTextDocument(document, { preview: false });
-        } catch (err) {
-          const reason = err instanceof Error ? err.message : String(err);
-          void vscode.window.showErrorMessage(`Failed to open link: ${reason}`);
-        }
+        await handleOpenMarkdownLink(message);
         break;
       }
       case 'openWorkspaceDiff': {
-        const p = message.path;
-        if (!p) break;
-        if (typeof message.oldContent !== 'string' || typeof message.newContent !== 'string') {
-          void vscode.window.showErrorMessage('Failed to open diff: missing diff content.');
-          break;
-        }
-        try {
-          let oldContent = message.oldContent;
-          let newContent = message.newContent;
-
-          if (message.preferGitHead === true) {
-            const wsRoot = getEffectiveWorkspaceRoot();
-            const { resolvedPath } = resolveWorkspaceFilePath(p);
-            const resolved = await resolveGitHeadDiffContents({
-              workspaceRoot: wsRoot,
-              resolvedPath,
-              fallbackOldContent: '',
-              fallbackNewContent: newContent,
-              execFileText,
-              readFileText: (filePath) => fs.readFile(filePath, 'utf8'),
-            });
-            oldContent = resolved.oldContent;
-            newContent = resolved.newContent;
-          }
-
-          await showWorkspaceDiff({ context, filePath: p, oldContent, newContent });
-        } catch (err) {
-          const reason = err instanceof Error ? err.message : String(err);
-          void vscode.window.showErrorMessage(`Failed to open diff: ${reason}`);
-        }
+        await handleOpenWorkspaceDiff({ context, message });
         break;
       }
       case 'requestHistory': {
-        await sendHistoryList();
+        await handleRequestHistory({ deps, host, settingsMgr, outputChannel, historyTitleGenerationInFlight });
         break;
       }
       case 'deleteConversation': {
-        const id = message.id;
-        if (!id) break;
-        const activeConversationId = conversation?.getConversationId?.();
-        if (activeConversationId && activeConversationId === id) {
-          void vscode.window.showWarningMessage('Cannot delete the active conversation.');
-          break;
-        }
-        try {
-          if (!/^[a-zA-Z0-9_-]+$/.test(id)) {
-            throw new Error('Invalid conversation id');
-          }
-          const convRoot = deps.getConversationStoreRoot() ?? (await deps.resolveConversationStoreRoot());
-          const resolvedRoot = path.resolve(convRoot);
-          const targetDir = path.resolve(convRoot, id);
-          const relative = path.relative(resolvedRoot, targetDir);
-          if (!relative || relative.startsWith('..') || path.isAbsolute(relative)) {
-            throw new Error('Invalid conversation id');
-          }
-          await fs.rm(targetDir, { recursive: true, force: true });
-        } catch (err) {
-          const reason = err instanceof Error ? err.message : String(err);
-          outputChannel?.appendLine(`[history] Failed to delete ${id}: ${reason}`);
-          void vscode.window.showErrorMessage(`Failed to delete conversation: ${reason}`);
-        }
-        await sendHistoryList();
+        await handleDeleteConversation({ deps, outputChannel, conversation, message });
+        await handleRequestHistory({ deps, host, settingsMgr, outputChannel, historyTitleGenerationInFlight });
         break;
       }
       case 'restoreConversation': {
-        const id = message.id;
-        if (!id) break;
-        try {
-          const maybe = conversation?.restoreConversation?.(id);
-          void Promise.resolve(maybe).catch((err: unknown) => {
-            const reason = err instanceof Error ? err.message : String(err);
-            outputChannel?.appendLine(`[restore] ${reason}`);
-            void vscode.window.showErrorMessage(`Failed to restore conversation: ${reason}`);
-          });
-        } catch (err) {
-          const reason = err instanceof Error ? err.message : String(err);
-          outputChannel?.appendLine(`[restore] ${reason}`);
-          void vscode.window.showErrorMessage(`Failed to restore conversation: ${reason}`);
-        }
+        handleRestoreConversation({ outputChannel, conversation, message });
         break;
       }
       case 'getConfig': {
@@ -717,638 +234,68 @@ export function createWebviewMessageHandler(deps: CreateWebviewMessageHandlerDep
         break;
       }
       case 'setLlmProfileId': {
-        const profileId = typeof message.profileId === 'string' ? message.profileId.trim() : '';
-        const currentSettings = await settingsMgr.get();
-        const previousProfileId = currentSettings.llm.profileId || '(none)';
-        const debugProfiles = currentSettings.agent?.debug === true;
-        const convMode = deps.getConversationMode();
-        const convStatus = conversation?.getStatus() ?? 'offline';
-        const convId = conversation?.getConversationId?.() || '(no conversation)';
-
-        if (debugProfiles) {
-          outputChannel?.appendLine('[LLM Profile] ========== SWITCH REQUESTED ==========');
-          outputChannel?.appendLine(
-            `[LLM Profile] Switch requested: ${previousProfileId} -> ${profileId || '(none)'} (mode=${convMode}, status=${convStatus}, id=${convId})`,
-          );
-          outputChannel?.appendLine('[LLM Profile] Current settings.llm: ' + JSON.stringify(currentSettings.llm));
-        }
-
-        try {
-          if (debugProfiles) {
-            outputChannel?.appendLine(
-              '[LLM Profile] Calling settingsMgr.update({ llm: { profileId: "' + profileId + '" } }, "global")...',
-            );
-          }
-          await settingsMgr.update({ llm: { profileId } }, 'global');
-          if (debugProfiles) {
-            outputChannel?.appendLine('[LLM Profile] Settings persisted successfully');
-          }
-        } catch (err) {
-          const reason = err instanceof Error ? err.message : String(err);
-          outputChannel?.appendLine('[LLM Profile] Failed to persist selection: ' + reason);
-          postStatusError(`Failed to save profile selection: ${reason}`);
-          break;
-        }
-
-        const updated = await settingsMgr.get();
-        if (debugProfiles) {
-          outputChannel?.appendLine('[LLM Profile] Updated settings.llm: ' + JSON.stringify(updated.llm));
-
-          outputChannel?.appendLine('[LLM Profile] Applying to conversation...');
-          outputChannel?.appendLine('[LLM Profile] conversation exists: ' + (conversation ? 'yes' : 'no'));
-          if (conversation) {
-            outputChannel?.appendLine(
-              '[LLM Profile] conversation.setSettings exists: ' + (typeof conversation.setSettings === 'function' ? 'yes' : 'no'),
-            );
-          }
-        }
-
-        try {
-          conversation?.setSettings(updated);
-          if (debugProfiles) {
-            outputChannel?.appendLine('[LLM Profile] Applied to conversation successfully');
-          }
-        } catch (err) {
-          const reason = err instanceof Error ? err.message : String(err);
-          outputChannel?.appendLine('[LLM Profile] Failed to apply to conversation: ' + reason);
-        }
-
-        const newLabel = resolveConfiguredLlmLabel(updated);
-        const oldLabel = deps.getLastKnownLlmLabel();
-        if (debugProfiles) {
-          outputChannel?.appendLine('[LLM Profile] Label: ' + (oldLabel || '(null)') + ' -> ' + (newLabel || '(null)'));
-        }
-        deps.setLastKnownLlmLabel(newLabel);
-
-        const finalStatus = conversation?.getStatus() ?? 'offline';
-        const finalMode = deps.getConversationMode();
-        const finalLabel = deps.getLastKnownLlmLabel();
-        if (debugProfiles) {
-          outputChannel?.appendLine(
-            '[LLM Profile] Posting status message: status=' + finalStatus + ', mode=' + finalMode + ', label=' + (finalLabel || '(null)'),
-          );
-        }
-
-        void host.postMessage({
-          type: 'status',
-          status: finalStatus,
-          mode: finalMode,
-          llmProfileLabel: finalLabel,
-        });
-
-        const profiles = listAvailableLlmProfiles();
-        const activeProfileId = updated.llm.profileId ?? null;
-        if (debugProfiles) {
-          outputChannel?.appendLine(
-            '[LLM Profile] Posting llmProfilesUpdated: profiles=[' + profiles.join(', ') + '], activeProfileId=' + (activeProfileId || '(null)'),
-          );
-        }
-
-        void host.postMessage({
-          type: 'llmProfilesUpdated',
-          profiles,
-          activeProfileId,
-        });
-
-        if (finalMode === 'remote' && finalStatus === 'online') {
-          void host.postMessage({
-            type: 'statusMessage',
-            level: 'warn',
-            message: 'Remote mode: LLM profile changes apply when you start a new conversation. The current remote conversation will continue using its existing model.',
-            autoDismiss: true,
-            autoDismissDelay: 8000,
-          });
-        }
-
-        outputChannel?.appendLine(
-          `[LLM Profile] Switched: ${previousProfileId} -> ${activeProfileId || '(none)'} (mode=${finalMode}, status=${finalStatus}, label=${finalLabel || '(null)'})`,
-        );
-        if (debugProfiles) {
-          outputChannel?.appendLine('[LLM Profile] ========== SWITCH COMPLETE ==========');
-        }
+        await handleSetLlmProfileId({ deps, host, settingsMgr, outputChannel, conversation, postStatusError, message });
         break;
       }
       case 'llmProfilesListRequest': {
-        const requestId = typeof message.requestId === 'string' ? message.requestId.trim() : '';
-        if (!requestId) break;
-        try {
-          const profiles = listAvailableLlmProfiles();
-          void host.postMessage({ type: 'llmProfilesListResponse', requestId, ok: true, profiles });
-        } catch (err) {
-          const reason = err instanceof Error ? err.message : String(err);
-          void host.postMessage({ type: 'llmProfilesListResponse', requestId, ok: false, error: reason });
-        }
+        handleLlmProfilesListRequest({ deps, host, outputChannel, message });
         break;
       }
       case 'llmProfileLoadRequest': {
-        const requestId = typeof message.requestId === 'string' ? message.requestId.trim() : '';
-        const profileId = typeof message.profileId === 'string' ? message.profileId.trim() : '';
-        if (!requestId || !profileId) break;
-
-        try {
-          const profile = llmProfilesStore.loadProfile(profileId, llmProfileStoreOptions());
-          void host.postMessage({
-            type: 'llmProfileLoadResponse',
-            requestId,
-            ok: true,
-            profileId: profile.profileId,
-            profile: profile.config,
-          });
-        } catch (err) {
-          const reason = err instanceof Error ? err.message : String(err);
-          void host.postMessage({ type: 'llmProfileLoadResponse', requestId, ok: false, profileId, error: reason });
-        }
+        handleLlmProfileLoadRequest({ deps, host, message });
         break;
       }
       case 'llmProfileSaveRequest': {
-        const requestId = typeof message.requestId === 'string' ? message.requestId.trim() : '';
-        const profileId = typeof message.profileId === 'string' ? message.profileId.trim() : '';
-        if (!requestId || !profileId) break;
-
-        try {
-          llmProfilesStore.saveProfile(profileId, message.profile, llmProfileStoreOptions());
-          void host.postMessage({ type: 'llmProfileSaveResponse', requestId, ok: true, profileId });
-
-          const updated = await settingsMgr.get();
-          void host.postMessage({
-            type: 'llmProfilesUpdated',
-            profiles: listAvailableLlmProfiles(),
-            activeProfileId: updated.llm.profileId ?? null,
-          });
-        } catch (err) {
-          const reason = err instanceof Error ? err.message : String(err);
-          void host.postMessage({ type: 'llmProfileSaveResponse', requestId, ok: false, profileId, error: reason });
-        }
+        await handleLlmProfileSaveRequest({ deps, host, settingsMgr, outputChannel, message });
         break;
       }
       case 'llmProfileDeleteRequest': {
-        const requestId = typeof message.requestId === 'string' ? message.requestId.trim() : '';
-        const profileId = typeof message.profileId === 'string' ? message.profileId.trim() : '';
-        if (!requestId || !profileId) break;
-
-        try {
-          llmProfilesStore.deleteProfile(profileId, llmProfileStoreOptions());
-          const key = getProfileApiKeySecretKey(profileId);
-          await context.secrets.delete(key);
-          deps.secretRegistry?.set(key, undefined);
-          void host.postMessage({ type: 'llmProfileDeleteResponse', requestId, ok: true, profileId });
-
-          const before = await settingsMgr.get();
-          const activeProfileId = before.llm.profileId ?? null;
-          if (activeProfileId === profileId) {
-            await settingsMgr.update({ llm: { profileId: '' } }, 'global');
-            void host.postMessage({
-              type: 'statusMessage',
-              level: 'error',
-              message: `Active LLM profile '${profileId}' was deleted; selection cleared.`,
-              autoDismiss: true,
-              autoDismissDelay: STATUS_MESSAGE_DISMISS_DELAY_MS,
-            });
-          } else {
-            void host.postMessage({
-              type: 'statusMessage',
-              level: 'info',
-              message: `Deleted profile '${profileId}'.`,
-              autoDismiss: true,
-              autoDismissDelay: STATUS_MESSAGE_DISMISS_DELAY_MS,
-            });
-          }
-
-          const updated = await settingsMgr.get();
-          deps.setLastKnownLlmLabel(resolveConfiguredLlmLabel(updated));
-
-          void host.postMessage({
-            type: 'status',
-            status: conversation?.getStatus() ?? 'offline',
-            mode: deps.getConversationMode(),
-            llmProfileLabel: deps.getLastKnownLlmLabel(),
-          });
-
-          void host.postMessage({
-            type: 'llmProfilesUpdated',
-            profiles: listAvailableLlmProfiles(),
-            activeProfileId: updated.llm.profileId ?? null,
-          });
-        } catch (err) {
-          const reason = err instanceof Error ? err.message : String(err);
-          void host.postMessage({ type: 'llmProfileDeleteResponse', requestId, ok: false, profileId, error: reason });
-        }
+        await handleLlmProfileDeleteRequest({ deps, host, context, settingsMgr, outputChannel, conversation, message });
         break;
       }
       case 'llmProfileApiKeyStatusRequest': {
-        const requestId = typeof message.requestId === 'string' ? message.requestId.trim() : '';
-        const profileId = typeof message.profileId === 'string' ? message.profileId.trim() : '';
-        if (!requestId || !profileId) break;
-
-        try {
-          const key = getProfileApiKeySecretKey(profileId);
-          const stored = await context.secrets.get(key);
-          const hasProfileKey = typeof stored === 'string' && stored.trim().length > 0;
-          const overrideProviderRaw = typeof message.provider === 'string' ? message.provider.trim() : '';
-          const overrideProvider = overrideProviderRaw && isLlmProvider(overrideProviderRaw) ? overrideProviderRaw : null;
-          const overrideBaseUrl = typeof message.baseUrl === 'string' ? message.baseUrl.trim() : '';
-
-          // Prefer explicit provider/baseUrl overrides so we can check provider keys even for
-          // draft/new profiles that do not exist yet on disk.
-          const provider = (() => {
-            if (overrideProvider) return overrideProvider;
-            if (overrideBaseUrl) return detectProviderFromBaseUrl(overrideBaseUrl);
-
-            const profile = llmProfilesStore.loadProfile(profileId, llmProfileStoreOptions());
-            return profile.config.provider ?? detectProviderFromBaseUrl(profile.config.baseUrl);
-          })();
-          const providerKeyName = getProviderApiKeyName(provider);
-          const hasProviderKey = await hasStoredSecret(providerKeyName);
-          void host.postMessage({
-            type: 'llmProfileApiKeyStatusResponse',
-            requestId,
-            ok: true,
-            profileId,
-            hasKey: hasProfileKey || hasProviderKey,
-            hasProfileKey,
-            hasProviderKey,
-            providerKeyName,
-          });
-        } catch (err) {
-          const reason = err instanceof Error ? err.message : String(err);
-          void host.postMessage({ type: 'llmProfileApiKeyStatusResponse', requestId, ok: false, profileId, error: reason });
-        }
+        await handleLlmProfileApiKeyStatusRequest({ deps, host, context, message });
         break;
       }
       case 'llmProfileApiKeySetRequest': {
-        const requestId = typeof message.requestId === 'string' ? message.requestId.trim() : '';
-        const profileId = typeof message.profileId === 'string' ? message.profileId.trim() : '';
-        const apiKey = typeof message.apiKey === 'string' ? message.apiKey.trim() : '';
-        if (!requestId || !profileId) break;
-
-        try {
-          const key = getProfileApiKeySecretKey(profileId);
-          if (!apiKey) {
-            await context.secrets.delete(key);
-            deps.secretRegistry?.set(key, undefined);
-          } else {
-            await context.secrets.store(key, apiKey);
-            deps.secretRegistry?.set(key, apiKey);
-          }
-          void host.postMessage({ type: 'llmProfileApiKeySetResponse', requestId, ok: true, profileId });
-        } catch (err) {
-          const reason = err instanceof Error ? err.message : String(err);
-          void host.postMessage({ type: 'llmProfileApiKeySetResponse', requestId, ok: false, profileId, error: reason });
-        }
+        await handleLlmProfileApiKeySetRequest({ deps, host, context, message });
         break;
       }
       case 'selectServer': {
-        const rawUrl = typeof message.url === 'string' ? message.url.trim() : '';
-        const url = rawUrl ? normalizeServerUrl(rawUrl) : { ok: true as const, url: '' };
-        if (!url.ok) {
-          postStatusError(url.error);
-          break;
-        }
-        const currentSettings = await settingsMgr.get();
-
-        const serverExists = currentSettings.servers.some((s) => s.url === url.url);
-        if (!serverExists && url.url) {
-          await settingsMgr.update({
-            servers: [...currentSettings.servers, { url: url.url }],
-            serverUrl: url.url,
-          });
-        } else {
-          await settingsMgr.update({ serverUrl: url.url });
-        }
-
-        const updated = await settingsMgr.get();
-        void host.postMessage({
-          type: 'serverListUpdated',
-          servers: updated.servers,
-          serverUrl: updated.serverUrl ?? '',
-        });
+        await handleSelectServer({ host, settingsMgr, postStatusError, message });
         break;
       }
       case 'addServer': {
-        const server = message.server;
-        if (!server?.url) break;
-
-        const normalized = normalizeServerUrl(server.url);
-        if (!normalized.ok) {
-          postStatusError(normalized.error);
-          break;
-        }
-
-        const label = typeof server.label === 'string' ? server.label.trim() : '';
-        const canonicalServer = label ? { url: normalized.url, label } : { url: normalized.url };
-
-        const currentSettings = await settingsMgr.get();
-        const exists = currentSettings.servers.some((s) => s.url === normalized.url);
-        if (!exists) {
-          const newServers = [...currentSettings.servers, canonicalServer];
-          await settingsMgr.update({ servers: newServers });
-          void host.postMessage({
-            type: 'serverListUpdated',
-            servers: newServers,
-            serverUrl: currentSettings.serverUrl ?? '',
-          });
-        }
+        await handleAddServer({ host, settingsMgr, postStatusError, message });
         break;
       }
       case 'removeServer': {
-        const rawUrl = typeof message.url === 'string' ? message.url.trim() : '';
-        if (!rawUrl) break;
-
-        const normalized = normalizeServerUrl(rawUrl);
-        if (!normalized.ok) {
-          postStatusError(normalized.error);
-          break;
-        }
-        const url = normalized.url;
-
-        const currentSettings = await settingsMgr.get();
-        const newServers = currentSettings.servers.filter((s) => s.url !== url);
-        const newServerUrl = currentSettings.serverUrl === url ? '' : currentSettings.serverUrl;
-
-        await settingsMgr.update({
-          servers: newServers,
-          serverUrl: newServerUrl,
-        });
-
-        void host.postMessage({
-          type: 'serverListUpdated',
-          servers: newServers,
-          serverUrl: newServerUrl ?? '',
-        });
+        await handleRemoveServer({ host, settingsMgr, postStatusError, message });
         break;
       }
       case 'switchToLocal': {
-        await settingsMgr.update({ serverUrl: '' });
-
-        const updated = await settingsMgr.get();
-        void host.postMessage({
-          type: 'serverListUpdated',
-          servers: updated.servers,
-          serverUrl: '',
-        });
+        await handleSwitchToLocal({ host, settingsMgr });
         break;
       }
       case 'selectAttachments': {
-        try {
-          const extensionMode = vscode.ExtensionMode;
-          const isTestMode =
-            extensionMode?.Test !== undefined &&
-            context.extensionMode === extensionMode.Test;
-          if (isTestMode && process.env.E2E_MOCK_ATTACHMENTS === '1') {
-            const mockUris = [vscode.Uri.joinPath(context.extensionUri, 'README.md')];
-            const attachments = await Promise.all(
-              mockUris.map(async (uri) => {
-                const label = toAttachmentLabel(uri);
-                let sizeBytes: number | undefined;
-                try {
-                  const stat = await vscode.workspace.fs.stat(uri);
-                  sizeBytes = stat.size;
-                } catch (err) {
-                  console.warn('[OpenHands] Failed to stat attachment', err);
-                }
-                return { uri: uri.toString(), label, sizeBytes };
-              })
-            );
-            void host.postMessage({ type: 'attachmentsSelected', attachments });
-            break;
-          }
-
-          const defaultUri = resolvePreferredWorkspaceFolderUri();
-          const picked = await vscode.window.showOpenDialog({
-            canSelectFiles: true,
-            canSelectFolders: false,
-            canSelectMany: true,
-            defaultUri,
-            openLabel: 'Attach',
-          });
-          if (!picked || picked.length === 0) break;
-
-          const attachments = await Promise.all(
-            picked.map(async (uri) => {
-              const label = toAttachmentLabel(uri);
-              let sizeBytes: number | undefined;
-              try {
-                const stat = await vscode.workspace.fs.stat(uri);
-                sizeBytes = stat.size;
-              } catch (err) {
-                console.warn('[OpenHands] Failed to stat attachment', err);
-              }
-              return { uri: uri.toString(), label, sizeBytes };
-            })
-          );
-          void host.postMessage({ type: 'attachmentsSelected', attachments });
-        } catch (err) {
-          const reason = err instanceof Error ? err.message : String(err);
-          void vscode.window.showErrorMessage(`Failed to select attachments: ${reason}`);
-        }
+        await handleSelectAttachments({ context, host, message });
         break;
       }
       case 'openAttachment': {
-        const raw = message.uri;
-        if (!raw) break;
-        try {
-          const uri = vscode.Uri.parse(raw, true);
-          const document = await vscode.workspace.openTextDocument(uri);
-          await vscode.window.showTextDocument(document, { preview: false });
-        } catch (err) {
-          const reason = err instanceof Error ? err.message : String(err);
-          void vscode.window.showErrorMessage(`Failed to open attachment: ${reason}`);
-        }
+        await handleOpenAttachment(message);
         break;
       }
       case 'send': {
         if (!conversation) break;
-        const baseText = message.text;
-        const contextFiles = Array.isArray(message.contextFiles)
-          ? message.contextFiles.filter((f): f is string => typeof f === 'string' && f.length > 0)
-          : [];
-        const attachmentUris = Array.isArray(message.attachments)
-          ? message.attachments
-            .filter((u): u is string => typeof u === 'string' && u.length > 0)
-            .map((u) => safeParseUri(u))
-            .filter((u): u is vscode.Uri => u !== undefined)
-          : [];
-
-        const globalStorageBaseDir = getGlobalStorageBaseDir(context.globalStorageUri?.fsPath);
-        const pastedImages = new Map<string, Uint8Array>();
-        const rewriteResult = rewriteDataImageMarkdown(baseText, (dataUrl) => {
-          const parsed = parseBase64DataImageUrl(dataUrl);
-          if (!parsed) return { url: '' };
-          if (parsed.bytes.length > MAX_PASTED_IMAGE_BYTES) return { url: '' };
-          pastedImages.set(parsed.imageId, parsed.bytes);
-          return { url: `${OPENHANDS_IMAGE_URL_PREFIX}${parsed.imageId}` };
-        });
-
-        let sanitizedText = rewriteResult.text;
-        if (pastedImages.size > 0) {
-          const failed = new Set<string>();
-          for (const [imageId, bytes] of pastedImages.entries()) {
-            try {
-              await persistPastedImage(globalStorageBaseDir, imageId, bytes);
-            } catch (err) {
-              failed.add(imageId);
-              const reason = err instanceof Error ? err.message : String(err);
-              outputChannel?.appendLine(`[pasted-images] Failed to persist ${imageId}: ${reason}`);
-            }
-          }
-          if (failed.size > 0) {
-            sanitizedText = rewriteOpenHandsImageUrls(sanitizedText, (imageId) => (failed.has(imageId) ? '' : undefined));
-            void vscode.window.showWarningMessage(`Some pasted images could not be saved (${failed.size}). They were omitted from the message.`);
-          }
-        } else if (rewriteResult.rewritten > 0) {
-          void vscode.window.showWarningMessage('Some pasted images were not supported and were omitted from the message.');
-        }
-
-        const attachmentsText = await buildAttachmentBlocks(attachmentUris);
-
-        let finalText = sanitizedText;
-        if (attachmentsText) {
-          finalText += attachmentsText;
-        }
-        if (contextFiles.length > 0) {
-          finalText += `\n\nUser has selected the following files for you to read:\n${contextFiles.join('\n')}`;
-        }
-
-        const queuedNotes = deps.getQueuedUserEditNotes();
-        const queuedNotesText = queuedNotes
-          .filter((note) => typeof note === 'string' && note.trim().length > 0)
-          .map((note) => note.trimEnd())
-          .join('\n\n');
-
-        if (queuedNotesText) {
-          await conversation.sendUserMessage(finalText, { extendedContent: [{ type: 'text', text: queuedNotesText }] });
-          deps.clearQueuedUserEditNotes();
-        } else {
-          await conversation.sendUserMessage(finalText);
-        }
+        await handleSend({ context, deps, conversation, message, outputChannel });
         break;
       }
       case 'halTtsRequest': {
-        if (typeof message.requestId !== 'string' || typeof message.conversationId !== 'string') break;
-        if (typeof message.stepIndex !== 'number' || !Number.isFinite(message.stepIndex)) break;
-
-        const stepIndex = Math.trunc(message.stepIndex);
-        if (stepIndex < 0) break;
-
-        const settings = await settingsMgr.get();
-        const script = getHalDialogueLinesForMode(settings.hal.userName, settings.hal.mode);
-        const line = script[stepIndex];
-        if (!line) {
-          void host.postMessage({ type: 'halTtsResponse', requestId: message.requestId, ok: false, error: 'Invalid HAL script line', shouldNotify: true });
-          break;
-        }
-
-        const apiKey = settings.secrets.halTtsApiKey ?? '';
-        const voiceId = line.voice === 'voice_hal' ? (settings.hal.voiceAId ?? '') : (settings.hal.voiceUserId ?? '');
-
-        const result = await getElevenlabsTtsGate().synthesize({
-          conversationId: message.conversationId,
-          apiKey,
-          voiceId,
-          text: line.text,
-          modelId: settings.hal.modelId,
-          cacheEnabled: settings.hal.cache,
-        });
-
-        if (result.ok) {
-          void host.postMessage({
-            type: 'halTtsResponse',
-            requestId: message.requestId,
-            ok: true,
-            audioBase64: Buffer.from(result.bytes).toString('base64'),
-            volume: settings.hal.volume,
-            mimeType: 'audio/mpeg',
-          });
-          break;
-        }
-
-        void host.postMessage({
-          type: 'halTtsResponse',
-          requestId: message.requestId,
-          ok: false,
-          error: result.error,
-          shouldNotify: result.shouldNotify,
-          disabled: result.disabled,
-        });
+        await handleHalTtsRequest({ deps, host, settingsMgr, getElevenlabsTtsGate, message });
         break;
       }
       case 'halVoiceConfirmRequest': {
-        if (typeof message.requestId !== 'string') break;
-        if (typeof message.mimeType !== 'string') break;
-        if (typeof message.audioBase64 !== 'string' || message.audioBase64.length === 0) {
-          void host.postMessage({ type: 'halVoiceConfirmResponse', requestId: message.requestId, ok: false, error: 'No audio provided' });
-          break;
-        }
-
-        const settings = await settingsMgr.get();
-
-        const defaultHalProfileId = DEFAULT_HAL_LLM_PROFILE_ID;
-        const configuredHalProfileId = typeof settings.hal.llmProfileId === 'string'
-          ? settings.hal.llmProfileId.trim()
-          : '';
-        const halProfileId = (() => {
-          if (!configuredHalProfileId) return defaultHalProfileId;
-          try {
-            validateProfileId(configuredHalProfileId);
-            return configuredHalProfileId;
-          } catch (err) {
-            const reason = err instanceof Error ? err.message : String(err);
-            outputChannel?.appendLine(`[hal] Invalid HAL llmProfileId '${configuredHalProfileId}'; using '${defaultHalProfileId}': ${reason}`);
-            return defaultHalProfileId;
-          }
-        })();
-
-        let halProfile: ReturnType<typeof llmProfilesStore.loadProfile> | undefined;
-        try {
-          halProfile = llmProfilesStore.loadProfile(halProfileId, llmProfileStoreOptions());
-        } catch (err) {
-          const reason = err instanceof Error ? err.message : String(err);
-          void host.postMessage({ type: 'halVoiceConfirmResponse', requestId: message.requestId, ok: false, error: reason });
-          outputChannel?.appendLine(`[hal] Failed to load HAL LLM profile '${halProfileId}': ${reason}`);
-          break;
-        }
-
-        const model = typeof halProfile.config.model === 'string' ? halProfile.config.model.trim() : '';
-        const baseUrlFromProfile = typeof halProfile.config.baseUrl === 'string' ? halProfile.config.baseUrl.trim() : '';
-        const provider = halProfile.config.provider ?? detectProviderFromBaseUrl(baseUrlFromProfile);
-        if (provider !== 'gemini') {
-          const error = `HAL voice_confirm requires a Gemini profile (got provider '${provider}').`;
-          void host.postMessage({ type: 'halVoiceConfirmResponse', requestId: message.requestId, ok: false, error });
-          outputChannel?.appendLine(`[hal] ${error}`);
-          break;
-        }
-
-        const baseUrl = baseUrlFromProfile || DEFAULT_PROVIDER_BASE_URLS.gemini;
-
-        const keyOrder = [
-          getProfileApiKeySecretKey(halProfileId),
-          getProviderApiKeyName(provider),
-          'openhands.llmApiKey',
-          'LLM_API_KEY',
-        ];
-        let halGeminiKey: string | undefined;
-        for (const key of keyOrder) {
-          const candidate = await getStoredSecret(key);
-          if (candidate) {
-            halGeminiKey = candidate;
-            break;
-          }
-        }
-
-        const result = await classifyHalVoiceDecision({
-          baseUrl,
-          apiKey: halGeminiKey ?? '',
-          model,
-          mimeType: message.mimeType,
-          audioBase64: message.audioBase64,
-        });
-
-        if (result.ok) {
-          void host.postMessage({ type: 'halVoiceConfirmResponse', requestId: message.requestId, ok: true, decision: result.decision });
-          break;
-        }
-
-        void host.postMessage({ type: 'halVoiceConfirmResponse', requestId: message.requestId, ok: false, error: result.error });
+        await handleHalVoiceConfirmRequest({ deps, host, context, settingsMgr, outputChannel, message });
         break;
       }
       case 'command': {
@@ -1404,67 +351,28 @@ export function createWebviewMessageHandler(deps: CreateWebviewMessageHandlerDep
         break;
       }
       case 'renderedEventsResponse':
-        deps.onRenderedEventsResponse(message.requestId, {
-          count: message.count,
-          eventTypes: message.eventTypes,
-          events: message.events,
-        });
+        handleRenderedEventsResponse({ deps, message });
         break;
       case 'uiStateResponse':
-        deps.onUiStateResponse(message.requestId, {
-          input: message.input,
-          showContextPicker: message.showContextPicker,
-          showSkillsPopover: message.showSkillsPopover,
-          showHistory: message.showHistory,
-          workspaceFilesCount: message.workspaceFilesCount,
-          selectedContextFiles: message.selectedContextFiles,
-          skillsCount: message.skillsCount,
-          attachmentsCount: message.attachmentsCount,
-          hasWelcomeProviderKey: message.hasWelcomeProviderKey,
-          hasWelcomeGeminiKey: message.hasWelcomeGeminiKey,
-          showWelcomeProviderKeyMessage: message.showWelcomeProviderKeyMessage,
-          showWelcomeGeminiKeyMessage: message.showWelcomeGeminiKeyMessage,
-        });
+        handleUiStateResponse({ deps, message });
         break;
       case 'halStateResponse':
-        deps.onHalStateResponse(message.requestId, {
-          enabled: message.enabled,
-          mode: message.mode,
-          phase: message.phase,
-          eye: message.eye,
-          stepIndex: message.stepIndex,
-          decision: message.decision,
-          lastError: message.lastError,
-        });
+        handleHalStateResponse({ deps, message });
         break;
       case 'webviewConsole': {
-        if (!deps.isDevBridgeEnabled()) break;
-        outputChannel?.appendLine(`[webview ${message.level}] ${message.args.join(' ')}`);
-        deps.fileLog(`[console.${message.level}] ${message.args.join(' ')}`);
+        handleWebviewConsole({ deps, outputChannel, message });
         break;
       }
       case 'webviewError': {
-        if (!deps.isDevBridgeEnabled()) break;
-        outputChannel?.appendLine(`[webview error] ${message.message}`);
-        if (message.stack) outputChannel?.appendLine(message.stack);
-        deps.fileLog(`[error] ${message.message}${message.stack ? `\n${message.stack}` : ''}`);
+        handleWebviewError({ deps, outputChannel, message });
         break;
       }
       case 'webviewNetwork': {
-        if (!deps.isDevBridgeEnabled()) break;
-        const line = `[webview net] ${message.phase} id=${message.id} ${message.method} ${message.url}${message.status !== undefined ? ` status=${message.status} ok=${message.ok}` : ''}`;
-        outputChannel?.appendLine(line);
-        deps.fileLog(line);
+        handleWebviewNetwork({ deps, outputChannel, message });
         break;
       }
       case 'webviewWebSocket': {
-        if (!deps.isDevBridgeEnabled()) break;
-        const parts = [`[webview ws] ${message.phase}`];
-        if (message.url) parts.push(`url=${message.url}`);
-        if (message.code !== undefined) parts.push(`code=${message.code}`);
-        if (message.reason) parts.push(`reason=${message.reason}`);
-        outputChannel?.appendLine(parts.join(' '));
-        deps.fileLog(parts.join(' '));
+        handleWebviewWebSocket({ deps, outputChannel, message });
         break;
       }
     }
